+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### 0.9.1
  - Add custom full topic support
+ - Multiple outputs ability for the Sensor node
  - Minor bugs fixed
  
 #### 0.9.0